feat: Rewrite 9 trivial coreutils in Rust + enable userspace clippy
Replace true, false, echo, pwd, sleep, rm, mkdir, rmdir, mv with
no_std Rust implementations in a new coreutils-rs multi-binary crate,
eliminating the musl-gcc dependency for these utilities. Binary sizes
drop from ~20KB+ (musl static) to 1.8–5KB (Rust no_std).

Also adds clippy -D warnings for all userspace crates to cargo xtask
check, fixing pre-existing warnings across syscall-lib, shell, edit,
init, passwd, su, id, and pty-test.

Key changes:
- New userspace/coreutils-rs crate with 9 [[bin]] targets
- syscall-lib: add mkdir(), rmdir(), unlink(), rename() wrappers
- syscall-lib: add # Safety docs, fix empty_loop, needless_range_loop
- xtask: move 9 utilities from C musl build to Rust userspace build
- xtask: add userspace clippy check covering all 16 userspace packages

+++ userspace/shell/src/main.rs
@@ -192,7 +192,7 @@ fn execute_pipeline(left: &[u8], right: &[u8]) {
     if left_pid == 0 {
         // Left child: stdout → write end of pipe.
         close(read_fd);
-        dup2(write_fd, STDOUT_FILENO as i32);
+        dup2(write_fd, STDOUT_FILENO);
         close(write_fd);
         exec_simple_command(left);
         exit(127);
@@ -209,7 +209,7 @@ fn execute_pipeline(left: &[u8], right: &[u8]) {
     if right_pid == 0 {
         // Right child: stdin → read end of pipe.
         close(write_fd);
-        dup2(read_fd, STDIN_FILENO as i32);
+        dup2(read_fd, STDIN_FILENO);
         close(read_fd);
         exec_simple_command(right);
         exit(127);
@@ -269,7 +269,7 @@ fn execute_external(
                 write_str(STDERR_FILENO, "sh: cannot open output file\n");
                 exit(1);
             }
-            dup2(fd as i32, STDOUT_FILENO as i32);
+            dup2(fd as i32, STDOUT_FILENO);
             close(fd as i32);
         }
 
@@ -283,7 +283,7 @@ fn execute_external(
                 write_str(STDERR_FILENO, "sh: cannot open input file\n");
                 exit(1);
             }
-            dup2(fd as i32, STDIN_FILENO as i32);
+            dup2(fd as i32, STDIN_FILENO);
             close(fd as i32);
         }
